requirements.txt
Django
djangorestframework
markdown
django-filter
mozilla-django-oidc
python-dotenv